fixing issue where setting outputPath, used by caching writer
